Update to latest ts-lib v0.3.0

+++ parserdialog.h
@@ -8,8 +8,8 @@
 
 // Project files
 #include "qhexedit.h"
-#include "public/mpeg2ts.h"
-#include "public/TsUtilities.h"
+#include "mpeg2ts/mpeg2ts.h"
+#include "mpeg2ts/TsUtilities.h"
 
 // Forward declarations
 namespace Ui {
@@ -52,7 +52,7 @@ private:
     tsutil::TsUtilities _tsUtil;
     std::vector<uint16_t> _pmtPids;
     mpeg2ts::PmtTable _pmt;
-    std::map<uint16_t, mpeg2ts::PmtTable> _pmtTables;
+    std::map<int, mpeg2ts::PmtTable> _pmtTables;
     mpeg2ts::PatTable _pat;
 };
